requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
python-dotenv>=1.0
SQLAlchemy>=2.0
requests>=2.31
apscheduler>=3.10
sentence-transformers>=2.2
scikit-learn>=1.3
numpy>=1.24
pydantic>=2.0
